# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActionErrorResponse": {
        "properties": {
          "code": {
            "enum": [
              "pending_action_not_found",
              "recommendation_not_found",
              "sku_not_in_recommendation",
              "invalid_quantity",
              "invalid_updated_fields",
              "version_conflict",
              "action_resolution_conflict",
              "action_expired",
              "catalog_not_found",
              "catalog_identifier_ambiguous",
              "sku_ambiguous",
              "catalog_identity_changed",
              "product_inactive",
              "sku_inactive",
              "insufficient_inventory",
              "cart_quantity_limit",
              "unsupported_action_schema",
              "invalid_action_payload",
              "expected_version_required"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/PendingActionErrorDetails"
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ActionErrorResponse",
        "type": "object"
      },
      "AddToCartPendingActionRequest": {
        "additionalProperties": false,
        "properties": {
          "quantity": {
            "default": 1,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "sku_id": {
            "format": "uuid",
            "title": "Sku Id",
            "type": "string"
          },
          "source_run_id": {
            "minLength": 1,
            "title": "Source Run Id",
            "type": "string"
          },
          "thread_id": {
            "minLength": 1,
            "title": "Thread Id",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "thread_id",
          "source_run_id",
          "sku_id"
        ],
        "title": "AddToCartPendingActionRequest",
        "type": "object"
      },
...
      "CatalogBrowseAddToCartPendingActionRequest": {
        "additionalProperties": false,
        "description": "Prepare a canonical SKU action from the read-only Catalog browse flow.",
        "properties": {
          "quantity": {
            "default": 1,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "sku_id": {
            "format": "uuid",
            "title": "Sku Id",
            "type": "string"
          },
          "thread_id": {
            "minLength": 1,
            "title": "Thread Id",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "thread_id",
          "sku_id"
        ],
        "title": "CatalogBrowseAddToCartPendingActionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PendingActionCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "thread_id": {
            "minLength": 1,
            "title": "Thread Id",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "thread_id",
          "expected_version"
        ],
        "title": "PendingActionCancelRequest",
        "type": "object"
      },
...
      "PendingActionTransitionResponse": {
        "properties": {
          "cart_item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CartItemView"
              },
              {
                "type": "null"
              }
            ]
          },
          "cart_quantity": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cart Quantity"
          },
          "current_money": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "pending_action": {
            "$ref": "#/components/schemas/PendingActionView"
          },
          "price_changed": {
            "default": false,
            "title": "Price Changed",
            "type": "boolean"
          },
          "requested_quantity": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Quantity"
          },
          "snapshot_money": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "pending_action"
        ],
        "title": "PendingActionTransitionResponse",
        "type": "object"
      },
      "PendingActionView": {
        "properties": {
          "action_type": {
            "enum": [
              "add_to_cart",
              "save_preference"
            ],
            "title": "Action Type",
            "type": "string"
          },
          "cancel_label": {
            "default": "Cancel",
            "title": "Cancel Label",
            "type": "string"
          },
          "confirm_label": {
            "default": "Confirm",
            "title": "Confirm Label",
            "type": "string"
          },
          "editable_fields": {
            "items": {
              "discriminator": {
                "mapping": {
                  "enum": "#/components/schemas/EnumEditableField",
                  "integer": "#/components/schemas/IntegerEditableField",
                  "text": "#/components/schemas/TextEditableField"
                },
                "propertyName": "field_type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/IntegerEditableField"
                },
                {
                  "$ref": "#/components/schemas/EnumEditableField"
                },
                {
                  "$ref": "#/components/schemas/TextEditableField"
                }
              ]
            },
            "title": "Editable Fields",
            "type": "array"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "pending_action_id": {
            "title": "Pending Action Id",
            "type": "string"
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AddToCartPreview"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview"
          },
          "risk_class": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Class",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "confirmed",
              "cancelled",
              "expired",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "pending_action_id",
          "action_type",
          "risk_class",
          "status",
          "version",
          "expires_at"
        ],
        "title": "PendingActionView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "FastAPI backend skeleton for the RetailPilot / ShopMind project.",
    "title": "RetailPilot Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/pending-actions/add-to-cart": {
      "post": {
        "operationId": "create_pending_action_api_pending_actions_add_to_cart_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ShopMind-Authenticated-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Authenticated-User"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartPendingActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pending Action",
        "tags": [
          "pending-actions"
        ]
      }
    },
    "/api/pending-actions/catalog-add-to-cart": {
      "post": {
        "operationId": "create_catalog_browse_pending_action_endpoint_api_pending_actions_catalog_add_to_cart_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ShopMind-Authenticated-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Authenticated-User"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogBrowseAddToCartPendingActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Catalog Browse Pending Action Endpoint",
        "tags": [
          "pending-actions"
        ]
      }
    },
    "/api/pending-actions/{pending_action_id}": {
      "get": {
        "operationId": "read_pending_action_api_pending_actions__pending_action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pending_action_id",
            "required": true,
            "schema": {
              "title": "Pending Action Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "thread_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Authenticated-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Authenticated-User"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Pending Action",
        "tags": [
          "pending-actions"
        ]
      }
    },
    "/api/pending-actions/{pending_action_id}/cancel": {
      "post": {
        "operationId": "cancel_pending_action_endpoint_api_pending_actions__pending_action_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "pending_action_id",
            "required": true,
            "schema": {
              "title": "Pending Action Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Authenticated-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Authenticated-User"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PendingActionCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionTransitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Pending Action Endpoint",
        "tags": [
          "pending-actions"
        ]
      }
    },
    "/api/pending-actions/{pending_action_id}/confirm": {
      "post": {
        "operationId": "confirm_pending_action_endpoint_api_pending_actions__pending_action_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "pending_action_id",
            "required": true,
            "schema": {
              "title": "Pending Action Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Authenticated-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Authenticated-User"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-ShopMind-Identity-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shopmind-Identity-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PendingActionTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionTransitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Pending Action Endpoint",
        "tags": [
          "pending-actions"
        ]
      }
    }
  }
}